+++ modules/hype-track.js
@@ -4,22 +4,23 @@ import * as Playback from "./playback.js";
 export default class HypeTrack {
     constructor() {
         this.playlist = null;
+        this.pausedSounds = [];
     }
 
     /**
      * Checks for the presence of the Hype Tracks playlist, creates one if none exist
      */
-    _checkForHypeTracksPlaylist() {
+    async _checkForHypeTracksPlaylist() {
         const enabled = game.settings.get(MAESTRO.MODULE_NAME, MAESTRO.SETTINGS_KEYS.HypeTrack.enable);
-        if(!game.user.isGM || !enabled) {
+        if(!enabled) {
             return;
         } 
 
         const hypePlaylist = game.playlists.entities.find(p => p.name == MAESTRO.DEFAULT_CONFIG.HypeTrack.playlistName);
-        if(!hypePlaylist) {
-            this.playlist = this._createHypeTracksPlaylist(true);
+        if(!hypePlaylist && game.user.isGM) {
+            this.playlist = await this._createHypeTracksPlaylist(true);
         } else {
-            this.playlist = hypePlaylist;
+            this.playlist = hypePlaylist || null;
         }
     }
 
@@ -40,18 +41,52 @@ export default class HypeTrack {
      * @param {Object} combat - the combat instance
      * @param {*} update - the update data
      */
-    async _checkHype(combat, update) {
-        if (!game.user.isGM || !Number.isNumeric(update.turn) || !combat.combatants.length) {
+    async _processHype(combat, update) {
+        if (!Number.isNumeric(update.turn) || !combat.combatants.length || !this.playlist) {
             return;
         }
-        const actorTrack = await this._getActorHypeTrack(combat.combatant.actor);
 
-        if (!actorTrack) {
-            return this.playlist.stopAll();
+        // Stop any active hype tracks
+        if (game.user.isGM && this?.playlist?.playing) {
+            this.playlist.stopAll();
         }
 
-        await this.playlist.stopAll();
-        Playback.playTrack(actorTrack, this.playlist._id);
+        // Find the hype track
+        const hypeTrack = this._getActorHypeTrack(combat.combatant.actor);
+        const pauseOthers = game.settings.get(MAESTRO.MODULE_NAME, MAESTRO.SETTINGS_KEYS.HypeTrack.pauseOthers);
+
+        if (!hypeTrack) {
+            if (this.pausedSounds.length) {
+                // Resume any previously paused sounds
+                Playback.resumeSounds(this.pausedSounds);
+                this.pausedSounds = [];
+            }
+            
+            return;
+        }
+
+        if (pauseOthers) {
+            // pause active playlists
+            this.pausedSounds = Playback.pauseAll();
+        }
+        
+
+        // Find the hype track's playlist sound and play it
+        const hypeTrackSound = this.playlist.sounds.find(s => s._id === hypeTrack);
+
+        if (game.user.isGM) {
+            await this.playHype(combat.combatant.actor, {warn: false});
+        }
+        
+        const howl = game.audio.sounds[hypeTrackSound.path].howl;
+
+        if (this.pausedSounds.length) {
+                // Defer the resumption of paused sounds after hype track finishes
+            howl.on("end", () => {
+                Playback.resumeSounds(this.pausedSounds);
+                this.pausedSounds = [];
+            });
+        }
     }
     
 
@@ -60,11 +95,11 @@ export default class HypeTrack {
      * @param {*} actor
      * 
      */
-    async _getActorHypeTrack(actor) {
+    _getActorHypeTrack(actor) {
         let actorTrack;
 
         try {
-            actorTrack = await actor.getFlag(MAESTRO.MODULE_NAME, MAESTRO.DEFAULT_CONFIG.HypeTrack.flagNames.track);
+            actorTrack = actor.getFlag(MAESTRO.MODULE_NAME, MAESTRO.DEFAULT_CONFIG.HypeTrack.flagNames.track);
             return actorTrack;
         } catch (e) {
             console.log(e);
@@ -93,11 +128,16 @@ export default class HypeTrack {
      * @param {Object} data 
      */
     async _addHypeButton (app, html, data) {
+        if(!game.user.isGM && !app?.entity?.owner) {
+            return;
+        }
+
         const enabled = game.settings.get(MAESTRO.MODULE_NAME, MAESTRO.SETTINGS_KEYS.HypeTrack.enable);
 
         if (!enabled) {
             return;
         }
+
         /**
          * Hype Button html literal
          * @todo replace with a template instead
@@ -127,8 +167,8 @@ export default class HypeTrack {
         /**
          * Register a click listener that opens the Hype Track form
          */
-        hypeButton.click(async ev => {
-            const actorTrack = await this._getActorHypeTrack(app.entity);
+        hypeButton.click(ev => {
+            const actorTrack = this._getActorHypeTrack(app.entity);
             this._openTrackForm(app.entity, actorTrack, {closeOnSubmit: true});
         });
     }
@@ -145,6 +185,57 @@ export default class HypeTrack {
             "playlist": this.playlist
         }
         new HypeTrackActorForm(actor, data, options).render(true);
+    }
+
+    /**
+     * Plays a hype track for the provided actor
+     * @param {*} actor 
+     */
+    async playHype(actor, {warn=true, pauseOthers=false}={}) {
+        if (typeof(actor) === "string") {
+            actor = game.actors.getName(actor) || null;
+        } else if (actor instanceof Object) {
+            actor = game.actors.getName(actor.name) || null;
+        }
+
+        if (!actor) {
+            if (warn) ui.notifications.warn(game.i18n.localize("HYPE-TRACK.PlayHype.NoActor"));
+            return;
+        }
+
+        const hypeTrack = this._getActorHypeTrack(actor);
+
+        if (!hypeTrack) {
+            if (warn) ui.notifications.warn(game.i18n.localize("HYPE-TRACK.PlayHype.NoTrack"));
+            return;
+        }
+
+        const playlist = this.playlist || game.playlists.entities.find(p => p.name === MAESTRO.DEFAULT_CONFIG.HypeTrack.playlistName || p.sounds.find(s => s._id === hypeTrack)) || null;
+
+        if (!playlist) {
+            if (warn) ui.notifications.warn(game.i18n.localize("HYPE-TRACK.PlayHype.NoPlaylist"));
+        }
+
+        if (playlist.playing) {
+            await playlist.stopAll();
+        }
+
+        let pausedSounds = [];
+
+        if (pauseOthers) {
+            pausedSounds = Playback.pauseAll();
+        }
+
+        const playedTrack = await Playback.playTrack(hypeTrack, playlist.id);
+
+        if (pauseOthers && pausedSounds.length) {
+            const playlistSound = playlist.sounds.find(s => s._id === playedTrack._id);
+            const howl = game.audio.sounds[playlistSound.path].howl;
+
+            howl.on("end", () => Playback.resumeSounds(pausedSounds));
+        }
+
+        return playedTrack;
     }
 }